pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "supersee"
version = "0.1.0"
description = "Crypto-AML investigation platform for asset managers entering digital assets."
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [{ name = "rakinnefoote" }]
keywords = ["xrpl", "compliance", "langgraph", "agentic", "aml"]

dependencies = [
    "xrpl-py>=4.0.0",
    "langgraph>=0.4,<0.5",
    "langgraph-checkpoint-postgres>=2.0,<3.0",
    "langchain-anthropic>=0.3,<0.4",
    "fastapi>=0.115,<0.116",
    "uvicorn[standard]>=0.32,<0.33",
    "psycopg[binary,pool]>=3.2,<3.3",
    "pydantic>=2.0,<3.0",
    "pydantic-settings>=2.0,<3.0",
    "httpx>=0.27,<0.28",
    "typer>=0.12,<0.13",
    # Pin click<8.3 — 8.3.x breaks single-decl bool flag semantics
    # (--pretty makes the value None instead of True). typer 0.12 +
    # click 8.2.x is the known-good combination.
    "click>=8.1,<8.3",
    "tenacity>=9.0,<10.0",
    "jinja2>=3.0,<4.0",
    "python-multipart>=0.0.9,<0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0,<9.0",
    "pytest-asyncio>=0.24,<0.25",
    "pytest-postgresql>=6.0,<7.0",
    "ruff>=0.6",
    "mypy>=1.11",
]

[project.scripts]
supersee = "supersee.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["supersee"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP", "B", "SIM", "ASYNC"]
ignore = [
    "E501",
    # B008: typer.Argument(...) in function defaults is the canonical
    # typer/click pattern. The rule is mostly intended to catch mutable
    # defaults and FastAPI Depends() shenanigans.
    "B008",
    # ASYNC109: httpx accepts a `timeout=` parameter; that's the library's
    # canonical pattern. We don't use anyio cancel scopes.
    "ASYNC109",
    # ASYNC240: pathlib.Path is used at CLI startup for one-time fixture
    # reads; not in a hot path. anyio.Path would be overengineered.
    "ASYNC240",
    # SIM117: `async with pool.connection() as conn: async with conn.cursor() as cur:`
    # is the canonical psycopg pool pattern. The flattened form (with
    # comma-separated context managers) works but the nested form is
    # idiomatic and explicit.
    "SIM117",
]

[tool.ruff.lint.per-file-ignores]
"supersee/**/tests/*.py" = [
    "B011",
    # SIM105: explicit try/except asyncio.CancelledError/pass is clearer
    # than contextlib.suppress in test setup/teardown where the cancel is
    # part of the test's intent.
    "SIM105",
]

[tool.mypy]
python_version = "3.12"
strict = true
warn_unused_configs = true
disallow_untyped_defs = true
warn_redundant_casts = true
warn_unused_ignores = true
# Test files are exercised by pytest; mypy strict on test code adds friction
# (esp. (await cur.fetchone())[0] indexing patterns and dict-vs-TypedDict
# mismatches) without catching real bugs. Pytest is the source of truth for
# test correctness; mypy is the source of truth for source code.
exclude = [
    "tests/",
]

[[tool.mypy.overrides]]
module = ["xrpl.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["supersee", "evals"]
addopts = "-ra -q --strict-markers"
markers = [
    "e2e: end-to-end tests requiring docker compose stack",
    "slow: tests that take more than 1 second",
]
